backend/requirements.txt
fastapi==0.115.0
uvicorn==0.29.0
python-multipart==0.0.9
langchain==0.3.7
langchain-groq==0.2.1
langchain-community==0.3.5
sentence-transformers==3.0.1
faiss-cpu==1.13.2
pydantic==2.9.2
python-dotenv==1.0.1